+++ src/card/sets/y2012/dark_ascension.rs
@@ -106,7 +106,9 @@ pub(in crate::card::sets) static RAY_OF_REVELATION: CardRecord = CardRecord::new
     "Ray of Revelation",
     CardArt::new("d7e2c5a4-cf92-46bd-9033-8036436488cb", "Cliff Childs"),
     CardSet::DarkAscension,
-    CardRules::new_instant(mana_cost!("{1}{W}")).with_abilities(&[
+    CardRules::new_instant(mana_cost!("{1}{W}"))
+        .with_flashback(mana_cost!("{G}"))
+        .with_abilities(&[
         AbilityDef::spell(
             "Destroy target enchantment.",
             EffectDef::Destroy {
@@ -124,9 +126,8 @@ pub(in crate::card::sets) static RAY_OF_REVELATION: CardRecord = CardRecord::new
                 owner: None,
             },
         )]),
-        AbilityDef::not_implemented(
+        abilities::flashback(
             "Flashback {G} (You may cast this card from your graveyard for its flashback cost. Then exile it.)",
-            "Casting from the graveyard for a flashback cost is not implemented.",
         ),
     ]),
 );